+++ lib/cover.dart
@@ -8,3 +8,7 @@
 /// cover --help
 /// ```
 library cover;
+
+export 'src/cover_command_runner.dart';
+export 'src/models/coverage_result.dart';
+export 'src/services/coverage_service.dart';