+++ lib/src/models/coverage_result.dart
@@ -1,0 +1,11 @@
+import 'package:lcov_parser/lcov_parser.dart';
+
+class CoverageResult {
+  const CoverageResult({
+    required this.coverage,
+    required this.files,
+  });
+
+  final double coverage;
+  final List<Record> files;
+}